Output functions refatctored
Moved output functions in an extension of ParsableCommand.
