+++ Sources/ScoutCLT/Main/ParsableCommand+PrintColorise.swift
@@ -1,0 +1,19 @@
+//
+// Scout
+// Copyright (c) Alexis Bridoux 2020
+// MIT license, see LICENSE file for details
+
+import Foundation
+import ArgumentParser
+import Scout
+
+extension ParsableCommand {
+
+    /// Retrieve the color file to colorise the output if one is found
+    func getColorFile() throws -> ColorFile? {
+        let colorFileURL = FileManager.default.homeDirectoryForCurrentUser.appendingPathComponent(".scout/Colors.plist")
+        guard let data = try? Data(contentsOf: colorFileURL) else { return nil }
+
+        return try PropertyListDecoder().decode(ColorFile.self, from: data)
+    }
+}
